Make IngestRecords integration test names more specific
